feat(aws): adds EKS addons module
Adds a new Terraform module to manage EKS addons.

This module allows users to easily deploy and manage a list of EKS addons to an EKS cluster.
It includes resources for creating addons, setting versions, and resolving conflicts.

The module also incorporates tagging functionality and utilizes a time static resource for consistent tagging.